Twenty Sixth Commit
Added Forgot Password Front End Along With The Back End With Three
Verification Options. User Can Verify Themselves By Either Email, SMS
Or Voice Call. The Email Is Customised With CSS3 To Give A Really
Appealing Look And Adds To The Ethnic Beauty Of The App. Coherance Is
Taken Care Of And Drawables Are Revised. Setup Gmail To Send IMAP Mails
Hassle Free. Professional Email Is Set Up Too. The Login UI Is Updated
To Keep The Firebase ID Updated And The Registration UI Sends Push
Notification After Successful Registration. The Phone Number Is Updated
After The OTP Is Verified And The Entries & Status Is Stored In MySQL
Database.